rename file using camelCase
