+++ assets/src/application/ui/calendars/dateDisplay/BiggieCalendarDate/index.tsx
@@ -8,7 +8,7 @@ import {
 	TIME_ONLY_12H_SHORT_FORMAT,
 	WEEKDAY_ONLY_FULL_FORMAT,
 	YEAR_ONLY_LONG_FORMAT,
-} from '@appConstants/date-fns-formats';
+} from '@appConstants/dateFnsFormats';
 import { CalendarDateProps } from '../types';
 import './style.css';
